requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asciichartpy==1.5.25
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
cheap-repr==0.5.1
click==8.1.3
climage==0.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
download==0.3.5
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.39.2
frosch==0.1.9
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
joblib==1.2.0
kdtree==0.16
kiwisolver==1.4.4
libusb1==3.0.0
markdown-it-py==2.2.0
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.3.230304
pathtools==0.1.2
Pillow==9.4.0
plotext==5.2.8
plotly==5.13.1
pretty-traceback==2022.1018
prettytable==3.6.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pure-python-adb==0.3.0.dev0
pyaes==1.6.1
pyasn1==0.4.8
pyfiglet==0.8.post1
Pygments==2.14.0
pyparsing==3.0.9
pyperclip==1.8.2
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
qrcode==7.4.2
regex==2022.10.31
requests==2.28.2
rich==13.3.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.17.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stack-data==0.1.4
Telethon==1.27.0
tenacity==8.2.2
termcolor==2.2.0
threadpoolctl==3.1.0
tk==0.1.0
tqdm==4.65.0
types-pytz==2022.7.1.2
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
wcwidth==0.2.6
yapf==0.31.0
yarl==1.8.2
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asciichartpy==1.5.25
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
cheap-repr==0.5.1
click==8.1.3
climage==0.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
download==0.3.5
elastic-transport==8.4.0
elasticsearch==8.6.2
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.39.2
frosch==0.1.9
frozenlist==1.3.3
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==4.13.0
Jinja2==3.1.2
joblib==1.2.0
kdtree==0.16
kiwisolver==1.4.4
libusb1==3.0.0
linkify-it-py==2.0.0
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-exclude==1.0.2
msgpack==1.0.5
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.3.230304
pathtools==0.1.2
pick==2.2.0
Pillow==9.4.0
plotext==5.2.8
plotly==5.13.1
pretty-traceback==2022.1018
prettytable==3.6.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pure-python-adb==0.3.0.dev0
pyaes==1.6.1
pyasn1==0.4.8
pydash==6.0.2
pyfiglet==0.8.post1
Pygments==2.14.0
pyparsing==3.0.9
pyperclip==1.8.2
pypng==0.20220715.0
pysolr==3.9.0
python-dateutil==2.8.2
pythonping==1.1.4
pytz==2022.7.1
PyYAML==6.0
pyyaml_env_tag==0.1
qrcode==7.4.2
regex==2022.10.31
requests==2.28.2
rich==13.3.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.17.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4
stack-data==0.1.4
Telethon==1.27.0
tenacity==8.2.2
termcolor==2.2.0
textual==0.16.0
threadpoolctl==3.1.0
tk==0.1.0
tqdm==4.65.0
types-pytz==2022.7.1.2
typing==3.7.4.3
typing_extensions==4.5.0
uc-micro-py==1.0.1
uiautomator==1.0.2
urllib3==1.26.15
wandb==0.14.0
watchdog==3.0.0
wcwidth==0.2.6
xarray==2023.3.0
xmltodict==0.13.0
yapf==0.31.0
yarl==1.8.2
zipp==3.15.0
